extracted components
